Block Selection!!!
This took me a long time to get working a lot of math refreshing, but essentially it outlines the block that you are pointing at by using Ray Casting. Its not working optimally but it gives an idea of what it should be doing. Next I will being updating so it only highlights the closest block and then I will implement destroy and add block functionality.